Update install.md with fix to garbled graphics
The (Debian and as such) Yunohost ISO has GRUB set up to boot in `vga=` to a specific resolution, which on certain (e.g. TV) monitors causes the screen to be completely glitched out.

This is an easy fix, once you know it. And I have seen people on the forum give up installing Yunohost because no-one mentioned the solution.